+++ packages/components/badge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.9](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.8...@spark-ui/badge@0.1.9) (2023-06-02)
+
+**Note:** Version bump only for package @spark-ui/badge
+
+## [0.1.8](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.7...@spark-ui/badge@0.1.8) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/badge
+
 ## [0.1.7](https://github.com/adevinta/spark/compare/@spark-ui/badge@0.1.6...@spark-ui/badge@0.1.7) (2023-05-25)
 
 **Note:** Version bump only for package @spark-ui/badge