+++ packages/components/radio-group/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/adevinta/spark/compare/@spark-ui/radio-group@1.0.1...@spark-ui/radio-group@1.0.2) (2023-06-08)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## 1.0.1 (2023-06-02)
+
+**Note:** Version bump only for package @spark-ui/radio-group
+
+## [2.0.6](https://github.com/adevinta/spark/compare/@spark-ui/radio@2.0.5...@spark-ui/radio@2.0.6) (2023-06-02)
+
+**Note:** Version bump only for package @spark-ui/radio
+
+## [2.0.5](https://github.com/adevinta/spark/compare/@spark-ui/radio@2.0.4...@spark-ui/radio@2.0.5) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/radio
+
+## [2.0.4](https://github.com/adevinta/spark/compare/@spark-ui/radio@2.0.3...@spark-ui/radio@2.0.4) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/radio
+
 ## [2.0.3](https://github.com/adevinta/spark/compare/@spark-ui/radio@2.0.2...@spark-ui/radio@2.0.3) (2023-05-31)
 
 **Note:** Version bump only for package @spark-ui/radio