+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.7](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.11.6...@spark-ui/switch@1.11.7) (2023-06-08)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.11.6](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.11.5...@spark-ui/switch@1.11.6) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.11.5](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.11.4...@spark-ui/switch@1.11.5) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [1.11.4](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.11.3...@spark-ui/switch@1.11.4) (2023-06-01)
+
+**Note:** Version bump only for package @spark-ui/switch
+
 ## [1.11.3](https://github.com/adevinta/spark/compare/@spark-ui/switch@1.11.2...@spark-ui/switch@1.11.3) (2023-05-31)
 
 **Note:** Version bump only for package @spark-ui/switch